publish v0.2.11

+++ nav_sim_modules/actioner/hueristic_autonomous_actioner/autonomous.py
@@ -23,7 +23,7 @@ class HueristicNavigationStack():
         allowable_angle: float,
         allowable_norm: float,
         avoidance_size: int,
-        path_planning_count: float, 
+        path_planning_count: int, 
         resolution: float=RESOLUTION
     ) -> None:
